add file identifier

+++ stop.sh
@@ -1,2 +1,5 @@
 #!/bin/bash
-kill $(cat /home/patartimotius/applicationtesting/apigateway/pid.file)
+if [ -f /home/patartimotius/applicationtesting/apigateway/pid.file ]; then
+	kill $(cat /home/patartimotius/applicationtesting/apigateway/pid.file)
+	echo 'kill them...'
+fi
